move test_stack.py to pytest
